frontend(auth): split shared auth base/utils and wire redirect flow
