+++ frontend/src/context/authContextBase.js
@@ -1,0 +1,10 @@
+// frontend/src/context/authContextBase.js
+import { createContext, useContext } from "react";
+
+export const AuthContext = createContext(null);
+
+export function useAuth() {
+  const ctx = useContext(AuthContext);
+  if (!ctx) throw new Error("useAuth must be used inside AuthProvider");
+  return ctx;
+}
