+++ frontend/src/main.jsx
@@ -4,7 +4,7 @@ import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 
 import App from "./App.jsx";
-import { AuthProvider } from "./context/AuthContext.jsx";
+import { AuthProvider } from "./context/AuthContext.jsx"; // ✅ correct import
 import "./index.css";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
